TicketMasterAPI init. This commit I've created and begun the ticketmasterapi branch. To begin this branch I've made sure that through the enpoint in the eventcontroller the call to the ticketmaster discovery v2 api is working and sending information back depending on the given latitude and longitude. The information that is sent back is a list of JSON obejct which each represents an event with details.
TODO: Make sure the objects are readable in the frontend.
TODO: Clean up the JSON objects in the backend as of now they contain
alot of uninportant information.
TODO: Map the events on the frontend mapClean up the JSON objects in the
backend as of now they contain alot of uninportant information.
TODO: Map the events on the frontend map.